Split out adapt from the intent_service.
